requirements.txt
accelerate
bitsandbytes
diffusers
matplotlib
numpy
peft
Pillow
protobuf
safetensors
sentencepiece
torch
torchvision
transformers
